Audio Medic!
Added audio for each of the clips we have.

Also small changes to random scripts because my last push made damage values only int.

+++ Assets/Scripts/Player/WeaponS/PlayerBullet.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class PlayerBullet : MonoBehaviour
 {
-	public int _damageVal;
+	public float _damageVal;
 
 	private void OnCollisionEnter2D(Collision2D collision)
 	{
